Update document identity column type

+++ IspDataDb/dbo/Tables/Document.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Document] (
-    [IdNo]            TINYINT       IDENTITY (1, 1) NOT NULL,
+    [IdNo]            SMALLINT      IDENTITY (1, 1) NOT NULL,
     [DocumentCode]    VARCHAR (5)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DocumentName]    VARCHAR (50)  COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [DocumentNameAra] VARCHAR (50)  COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -14,4 +14,6 @@
     [CreateDate]      DATETIME2 (7) NULL,
     CONSTRAINT [PK_Document] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
